Moving all thing in place
