+++ lib/compositionhelper.js
