+++ lib/escapeSequence.js
