+++ lib/eventemitter.js
