+++ lib/utils/contains.js
