+++ lib/utils/dom/isBoldBroken.js
