+++ lib/utils/isThirdLevelShift.js
