+++ lib/utils/matchColor.js
