+++ src/addons/fit.js
